refactor: restructuring of queryable methods
move helper methods and classes to separate file, add provider queryables to core method and make it return dict of queryable properties instead of set of property names
