+++ eodag/rest/utils.py
@@ -33,22 +33,20 @@ from typing import (
     List,
     NamedTuple,
     Optional,
-    Set,
     Tuple,
     Union,
 )
 from urllib.parse import urlencode
 
 import dateutil.parser
-import requests
 from dateutil import tz
 from fastapi.responses import StreamingResponse
-from pydantic import BaseModel, Field
 from shapely.geometry import Polygon, shape
 
 import eodag
 from eodag import EOProduct
 from eodag.api.product.metadata_mapping import OSEO_METADATA_MAPPING
+from eodag.api.queryables import QueryableProperty
 from eodag.api.search_result import SearchResult
 from eodag.config import load_stac_config, load_stac_provider_config
 from eodag.plugins.crunch.filter_latest_intersect import FilterLatestIntersect
@@ -59,7 +57,6 @@ from eodag.utils import (
     DEFAULT_ITEMS_PER_PAGE,
     DEFAULT_PAGE,
     GENERIC_PRODUCT_TYPE,
-    USER_AGENT,
     _deprecated,
     dict_items_recursive_apply,
     string_to_jsonpath,
@@ -68,7 +65,6 @@ from eodag.utils.exceptions import (
     MisconfiguredError,
     NoMatchingProductType,
     NotAvailableError,
-    RequestError,
     UnsupportedProductType,
     ValidationError,
 )
@@ -1057,119 +1053,9 @@ def get_stac_extension_oseo(url: str) -> Dict[str, str]:
     )
 
 
-class QueryableProperty(BaseModel):
-    """A class representing a queryable property.
-
-    :param description: The description of the queryables property
-    :type description: str
-    :param ref: (optional) A reference link to the schema of the property.
-    :type ref: str
-    """
-
-    description: str
-    ref: Optional[str] = Field(default=None, serialization_alias="$ref")
-    type: Optional[list] = None
-
-
-class Queryables(BaseModel):
-    """A class representing queryable properties for the STAC API.
-
-    :param json_schema: The URL of the JSON schema.
-    :type json_schema: str
-    :param q_id: (optional) The identifier of the queryables.
-    :type q_id: str
-    :param q_type: The type of the object.
-    :type q_type: str
-    :param title: The title of the queryables.
-    :type title: str
-    :param description: The description of the queryables
-    :type description: str
-    :param properties: A dictionary of queryable properties.
-    :type properties: dict
-    :param additional_properties: Whether additional properties are allowed.
-    :type additional_properties: bool
-    """
-
-    json_schema: str = Field(
-        default="https://json-schema.org/draft/2019-09/schema",
-        serialization_alias="$schema",
-    )
-    q_id: Optional[str] = Field(default=None, serialization_alias="$id")
-    q_type: str = Field(default="object", serialization_alias="type")
-    title: str = Field(default="Queryables for EODAG STAC API")
-    description: str = Field(
-        default="Queryable names for the EODAG STAC API Item Search filter."
-    )
-    properties: Dict[str, QueryableProperty] = Field(
-        default={
-            "id": QueryableProperty(
-                description="ID",
-                ref="https://schemas.stacspec.org/v1.0.0/item-spec/json-schema/item.json#/id",
-            ),
-            "collection": QueryableProperty(
-                description="Collection",
-                ref="https://schemas.stacspec.org/v1.0.0/item-spec/json-schema/item.json#/collection",
-            ),
-            "geometry": QueryableProperty(
-                description="Geometry",
-                ref="https://schemas.stacspec.org/v1.0.0/item-spec/json-schema/item.json#/geometry",
-            ),
-            "bbox": QueryableProperty(
-                description="Bbox",
-                ref="https://schemas.stacspec.org/v1.0.0/item-spec/json-schema/item.json#/bbox",
-            ),
-            "datetime": QueryableProperty(
-                description="Datetime",
-                ref="https://schemas.stacspec.org/v1.0.0/item-spec/json-schema/datetime.json#/properties/datetime",
-            ),
-        }
-    )
-    additional_properties: bool = Field(
-        default=True, serialization_alias="additionalProperties"
-    )
-
-    def get_properties(self) -> Dict[str, QueryableProperty]:
-        """Get the queryable properties.
-
-        :returns: A dictionary containing queryable properties.
-        :rtype: typing.Dict[str, QueryableProperty]
-        """
-        return self.properties
-
-    def __contains__(self, name: str) -> bool:
-        return name in self.properties
-
-    def __setitem__(self, name: str, qprop: QueryableProperty) -> None:
-        self.properties[name] = qprop
-
-
-def rename_to_stac_standard(key: str) -> str:
-    """Fetch the queryable properties for a collection.
-
-    :param key: The camelCase key name obtained from a collection's metadata mapping.
-    :type key: str
-    :returns: The STAC-standardized property name if it exists, else the default camelCase queryable name
-    :rtype: str
-    """
-    # Load the stac config properties for renaming the properties
-    # to their STAC standard
-    stac_config_properties: Dict[str, Any] = stac_config["item"]["properties"]
-
-    for stac_property, value in stac_config_properties.items():
-        if isinstance(value, list):
-            value = value[0]
-        if str(value).endswith(key):
-            return stac_property
-
-    if key in OSEO_METADATA_MAPPING:
-        return "oseo:" + key
-
-    return key
-
-
 def fetch_collection_queryable_properties(
-    collection_id: str, provider: Optional[str] = None
-) -> Set[str]:
+    collection_id: Optional[str] = None, provider: Optional[str] = None
+) -> Dict[str, QueryableProperty]:
     """Fetch the queryable properties for a collection.
 
     :param collection_id: The ID of the collection.
@@ -1180,140 +1066,12 @@ def fetch_collection_queryable_properties(
     :rtype queryable_properties: set
     """
     # Fetch the metadata mapping for collection-specific queryables
-    kwargs = {"product_type": collection_id}
+    kwargs = {}
+    if collection_id is not None:
+        kwargs["product_type"] = collection_id
     if provider is not None:
         kwargs["provider"] = provider
-    eodag_queryable_properties = eodag_api.get_queryables(**kwargs)
-
-    # list of all the STAC standardized collection-specific queryables
-    queryable_properties: Set[str] = set()
-    for prop in eodag_queryable_properties:
-        # remove pure eodag properties
-        if prop not in ["start", "end", "geom", "locations", "id"]:
-            queryable_properties.add(rename_to_stac_standard(prop))
-    return queryable_properties
-
-
-def add_provider_queryables(provider: str, queryables: dict) -> Dict[str, Any]:
-    """Add the queryables fetched from the given provider to the default queryables.
-    If the queryables endpoint is not supported by the provider, the original queryables are returned
-    :param provider: The provider.
-    :type provider: str
-    :param queryables: default queryables to which provider queryables will be added
-    :type queryables: dict
-    :returns queryable_properties: A dict containing the formatted queryable properties
-                                   including queryables fetched from the provider.
-    :rtype dict
-    """
-    search_plugin = next(
-        eodag_api._plugins_manager.get_search_plugins(provider=provider)
-    )
-    search_type = search_plugin.config.type
-    if search_type == "StacSearch":
-        queryables_url = search_plugin.config.api_endpoint.replace(
-            "/search", "/queryables"
-        )
-    else:
-        queryables_url = ""
-    if not queryables_url:
-        logger.info("no url was found for %s provider-specific queryables", provider)
-        return queryables
-    try:
-        headers = USER_AGENT
-        if hasattr(search_plugin, "auth"):
-            res = requests.get(queryables_url, headers=headers, auth=search_plugin.auth)
-        else:
-            res = requests.get(queryables_url, headers=headers)
-        res.raise_for_status()
-    except requests.exceptions.HTTPError as err:
-        if err.response.status_code == 404:
-            logger.info("provider %s does not support queryables", provider)
-            return queryables
-        else:
-            raise RequestError(str(err))
-    else:
-        provider_queryables = res.json()["properties"]
-        return _format_provider_queryables(provider_queryables, queryables)
-
-
-def add_provider_product_type_queryables(
-    provider: str, product_type: str, queryables: dict
-) -> Dict[str, Any]:
-    """Add the queryables fetched from the given provider for the given product type
-    to the default queryables derived from the metadata mapping.
-    If the queryables endpoint is not supported by the provider, the original queryables are returned.
-    :param provider: The provider.
-    :type provider: str
-    :param product_type: EODAG product type
-    :type product_type: str
-    :param queryables: default queryables to which provider queryables will be added
-    :type queryables: dict
-    :returns queryable_properties: A dict containing the formatted queryable properties
-                                   including queryables fetched from the provider.
-    :rtype dict
-    """
-    search_plugin = next(
-        eodag_api._plugins_manager.get_search_plugins(provider=provider)
-    )
-    search_type = search_plugin.config.type
-    provider_product_type = search_plugin.config.products.get(product_type, {}).get(
-        "productType", None
-    )
-    if not provider_product_type:
-        logger.warning(
-            "provider product type mapping for product type %s not found", product_type
-        )
-        provider_product_type = product_type
-    if search_type == "StacSearch":
-        api_url = search_plugin.config.api_endpoint.replace("/search", "/")
-        queryables_url = (
-            api_url + "collections/" + provider_product_type + "/queryables"
-        )
-    else:
-        queryables_url = ""
-
-    if not queryables_url:
-        logger.info(
-            "no url was found for %s on %s provider-specific queryables",
-            product_type,
-            provider,
-        )
-        return queryables
-    try:
-        headers = USER_AGENT
-        if hasattr(search_plugin, "auth"):
-            res = requests.get(queryables_url, headers=headers, auth=search_plugin.auth)
-        else:
-            res = requests.get(queryables_url, headers=headers)
-        res.raise_for_status()
-    except requests.exceptions.HTTPError as err:
-        if err.response.status_code == 404:
-            logger.info("provider %s does not support queryables", provider)
-            return queryables
-        else:
-            raise RequestError(str(err))
-    else:
-        if "properties" in res.json():
-            provider_queryables = res.json()["properties"]
-            return _format_provider_queryables(provider_queryables, queryables)
-        else:
-            return queryables
-
-
-def _format_provider_queryables(
-    provider_queryables: dict, queryables: dict
-) -> Dict[str, Any]:
-    for queryable, data in provider_queryables.items():
-        attributes = {"description": queryable}
-        if "type" in data:
-            if isinstance(data["type"], list):
-                attributes["type"] = data["type"]
-            else:
-                attributes["type"] = [data["type"]]
-        if "ref" in data:
-            attributes["ref"] = data["ref"]
-        queryables[queryable] = QueryableProperty(**attributes)
-    return queryables
+    return eodag_api.get_queryables(**kwargs)
 
 
 def eodag_api_init() -> None:
